refactor(clickhouse sink): fold prepared state into config module
